+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.30.0] - 2026-04-09
+### Added
+- Added support for translation memories in text translation via
+  `translation_memory` and `translation_memory_threshold` parameters in
+  `translate_text()`.
+- Added `list_translation_memories()` method to the `DeepLClient` class for
+  listing available translation memories.
+- Added `--translation-memory-id` and `--translation-memory-threshold` CLI
+  arguments.
+
 ### Fixed
 - Fixed "gitlab release" CI job to explicitly set the release-cli image, avoiding issues with missing `$HOME` from the global image.
 
@@ -395,7 +406,8 @@ Version increased to avoid conflicts with old packages on PyPI.
 ## [0.1.0] - 2021-07-26
 Initial version.
 
-[Unreleased]: https://github.com/DeepLcom/deepl-python/compare/v1.29.0...HEAD
+[Unreleased]: https://github.com/DeepLcom/deepl-python/compare/v1.30.0...HEAD
+[1.30.0]: https://github.com/DeepLcom/deepl-python/compare/v1.29.0...v1.30.0
 [1.29.0]: https://github.com/DeepLcom/deepl-python/compare/v1.28.0...v1.29.0
 [1.28.0]: https://github.com/DeepLcom/deepl-python/compare/v1.27.0...v1.28.0
 [1.27.0]: https://github.com/DeepLcom/deepl-python/compare/v1.26.0...v1.27.0